fix: don't seek to the start when swiping right
The API has an extra parameter `allow_seeking` that indicates whether
the seek left is due to pressing the "prev" button, or because the album
is swiped right to go to the previous track immediately.

This patch adds support for this feature.

There's still one issue however: when going to the previous song, the
player will for a short time show the first track in an album. Not sure
why, I suspect go-librespot doesn't send the whole queue (so the player
doesn't know there are other tracks in the album).