fix: don't try to seek to the start of the song when starting from there
Apparently a bit of the song has already been decoded and sent to the
audio device before the seek. When starting a new song, this causes the
song to revert to time 0 causing a slight stutter.

The stutter is presumably still there when an actual seek is needed. But
that's probably a less common case. A real fix would be to only start
decoding after the seek (which doesn't seem to be the case here).