feat: improve efficiency of the decryptor
The decryptor was using up a lot of time. For a ~10min song it took
about twice the time that it took to run the Vorbis decoder and this
time would only go up considering the exponential nature of the loop.

But actually, all it did was count up a number. This can be done by a
simple addition instead of a complicated loop. Hence this patch changes
the loop to a simple addition.

It is 32-bits at the moment, which means it would overflow with a stream
that's 64GB big. This is entirely fixable if needed, but 64GB seems like
an unlikely large stream.