feat: refactor volume change events
This changes the way volume changes from Spotify, the HTTP API, or an
external mixer are reported back to Spotify.

  * Use a regular Go channel instead of a custom queue. Updates still
    use non-blocking sends.
  * Listen for volume changes in the AppPlayer instead of spinning up a
    new goroutine for this purpose. This might also fix some race
    conditions.
  * Rate limit all volume changes in the same way that the Spotify Web
    Player does: by sending the new volume after no change has been
    observed for 1 second.
    This seems to be good enough to avoid any HTTP status 429 errors.
    Importantly, this also rate limits volume updates from Spotify
    itself. Previously it was very easy to cause 429 errors by simply
    sliding the volume slider a lot.

I have tested this change with both the ALSA output and the PulseAudio
output. I have not tested this change with the external mixer configured
in ALSA, because I don't know how to do that on a PipeWire system.