fix: avoid popping noise when changing volume
Just a little pet peeve of mine. When the volume changes (or is
muted/unmuted), a little pop can be heard due to the sudden change. This
can be fixed by smoothly increasing/decreasing the volume.

This noise is especially audible on tracks like this one:
https://www.youtube.com/watch?v=eH-5U7GeofA

Weirdly, PulseAudio and PipeWire don't seem to do this internally. Maybe
I'll send a patch to PipeWire to fix this, because it's a bit annoying
(and there's not much we can do about it in the PulseAudio driver apart
from managing volume ourselves).