fix: fix fast relative volume updates
Relative volume changes would use old values, not the new value set by a
volume change in the last second. To fix this, update the player state
immediately and only delay sending the update (instead of also delaying
the update to the player state).

This could mean that another change (such as play/pause) would send an
earlier volume update, but I don't think that's a problem. We'll send a
PutStateReason_VOLUME_CHANGED anyway after a second.